+++ cmd/frontend/internal/app/updatecheck/handler.go
@@ -188,6 +188,7 @@ type pingRequest struct {
 	NewCodeIntelUsage    json.RawMessage `json:"newCodeIntelUsage"`
 	SearchUsage          json.RawMessage `json:"searchUsage"`
 	ExtensionsUsage      json.RawMessage `json:"extensionsUsage"`
+	CodeInsightsUsage    json.RawMessage `json:"codeInsightsUsage"`
 	InitialAdminEmail    string          `json:"initAdmin"`
 	TotalUsers           int32           `json:"totalUsers"`
 	HasRepos             bool            `json:"repos"`
@@ -290,6 +291,7 @@ type pingPayload struct {
 	SearchOnboarding     json.RawMessage `json:"search_onboarding"`
 	DependencyVersions   json.RawMessage `json:"dependency_versions"`
 	ExtensionsUsage      json.RawMessage `json:"extensions_usage"`
+	CodeInsightsUsage    json.RawMessage `json:"code_insights_usage"`
 	InstallerEmail       string          `json:"installer_email"`
 	AuthProviders        string          `json:"auth_providers"`
 	ExtServices          string          `json:"ext_services"`
@@ -373,6 +375,7 @@ func marshalPing(pr *pingRequest, hasUpdate bool, clientAddr string, now time.Ti
 		InstallerEmail:       pr.InitialAdminEmail,
 		DependencyVersions:   pr.DependencyVersions,
 		ExtensionsUsage:      pr.ExtensionsUsage,
+		CodeInsightsUsage:    pr.CodeInsightsUsage,
 		AuthProviders:        strings.Join(pr.AuthProviders, ","),
 		ExtServices:          strings.Join(pr.ExternalServices, ","),
 		BuiltinSignupAllowed: strconv.FormatBool(pr.BuiltinSignupAllowed),